chore: refactor rtk with type contracts
