+++ packages/core/content-releases/admin/src/services/release.ts
@@ -1,13 +1,10 @@
 import { createApi } from '@reduxjs/toolkit/query/react';
 
 import { pluginId } from '../pluginId';
-import { axiosBaseQuery } from '../utils/data';
 
-import type {
-  CreateRelease,
-  GetReleases,
-  ReleaseDataResponse,
-} from '../../../shared/contracts/releases';
+import { axiosBaseQuery } from './utils';
+
+import type { CreateRelease, GetReleases } from '../../../shared/contracts/releases';
 
 const releaseApi = createApi({
   reducerPath: pluginId,
@@ -24,7 +21,7 @@ const releaseApi = createApi({
         },
         providesTags: ['Releases'],
       }),
-      createRelease: build.mutation<{ data: ReleaseDataResponse }, CreateRelease.Request['body']>({
+      createRelease: build.mutation<CreateRelease.Response, CreateRelease.Request['body']>({
         query(data) {
           return {
             url: '/content-releases',
