+++ packages/core/content-releases/shared/contracts/releases.ts
@@ -31,15 +31,11 @@ export declare namespace GetReleases {
     query?: Partial<Pick<Pagination, 'page' | 'pageSize'>>;
   }
 
-  export type Response =
-    | {
-        data: ReleaseDataResponse[];
-        pagination: Pagination;
-      }
-    | {
-        data: null;
-        error: errors.ApplicationError;
-      };
+  export interface Response {
+    data: ReleaseDataResponse[];
+    error?: errors.ApplicationError;
+    pagination: Pagination;
+  }
 }
 
 /**
@@ -55,14 +51,10 @@ export declare namespace GetRelease {
     };
   }
 
-  export type Response =
-    | {
-        data: ReleaseDataResponse;
-      }
-    | {
-        data: null;
-        error: errors.ApplicationError | errors.NotFoundError;
-      };
+  export type Response = {
+    data: ReleaseDataResponse;
+    error?: errors.ApplicationError | errors.NotFoundError;
+  };
 }
 
 /**
@@ -78,9 +70,10 @@ export declare namespace CreateRelease {
     };
   }
 
-  export type Response =
-    | { data: ReleaseDataResponse }
-    | { data: null; error: errors.ApplicationError | errors.ValidationError };
+  export interface Response {
+    data: ReleaseDataResponse;
+    error?: errors.ApplicationError | errors.ValidationError;
+  }
 }
 
 /**
@@ -99,7 +92,8 @@ export declare namespace UpdateRelease {
     };
   }
 
-  export type Response =
-    | { data: ReleaseDataResponse }
-    | { data: null; error: errors.ApplicationError | errors.ValidationError };
+  export interface Response {
+    data: ReleaseDataResponse;
+    error?: errors.ApplicationError | errors.ValidationError;
+  }
 }
